Fixed compile error when building against jzmq

+++ src/main/java/org/zeromq/jzmq/poll/PollerImpl.java
@@ -46,7 +46,7 @@ public class PollerImpl implements Poller {
 
             // simulate ETERM to make JeroMQ act like jzmq, which no longer returns -1
             if (numberOfObjects < 0) {
-                throw new ZMQException("Simulated ETERM error", ZMQ.Error.ETERM.getCode());
+                throw new ZMQException("Simulated ETERM error", (int) ZMQ.Error.ETERM.getCode());
             }
         } catch (ZMQException ex) {
             throw ZMQExceptions.wrap(ex);
